pyproject.toml
[tool.poetry]
name = "partial-cli"
version = "0.1.0"
description = ""
authors = ["Karlkim Suwanmongkol <karlkim@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
rich-click = "^1.8.3"
asyncio = "^3.4.3"


[tool.poetry.group.chia.dependencies]
chia-blockchain = "^2.4.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
partial = "partial_cli:partial_cmd"